+++ tests/legs.rs
@@ -20,7 +20,6 @@ use metropolis::schedule_utility::alpha_beta_gamma::AlphaBetaGammaModel;
 use metropolis::schedule_utility::ScheduleUtility;
 use metropolis::simulation::results::AgentResult;
 use metropolis::simulation::Simulation;
-use metropolis::stop::StopCriterion;
 use metropolis::travel_utility::{PolynomialFunction, TravelUtility};
 use metropolis::units::{Flow, Interval, Lanes, Length, Speed, Time, Utility, ValueOfTime, PCE};
 use num_traits::Float;
@@ -157,7 +156,6 @@ fn get_simulation() -> Simulation<f64> {
         output_directory: Default::default(),
         period: Interval([Time(0.0), Time(50.0)]),
         learning_model: LearningModel::Exponential(0.0),
-        stopping_criteria: vec![StopCriterion::MaxIteration(1)],
         network: NetworkParameters {
             road_network: Some(RoadNetworkParameters {
                 contraction: Default::default(),
@@ -171,6 +169,7 @@ fn get_simulation() -> Simulation<f64> {
             }),
         },
         init_iteration_counter: 1,
+        max_iterations: 1,
         update_ratio: 1.0,
         random_seed: None,
         nb_threads: 0,